main.py runs on gpu if we ask for more memory, will refactor so we operate by command line

+++ dnn/pipeline.py
@@ -59,6 +59,6 @@ def read_tiff_gen(tiff_files, side):
 
                     img = np.stack(bands, axis=-1)
                     if (img != 0).any():
-                        yield img
+                        yield img.astype(np.float32)
 
     return gen
